Refactor_Front/Apply_Typescript_v2[ Paging apply typescript]
