+++ front/package.json
@@ -27,6 +27,7 @@
     "@testing-library/user-event": "^12.1.10",
     "axios": "^0.24.0",
     "date-fns": "^2.25.0",
+    "global": "^4.4.0",
     "immer": "^9.0.6",
     "material-ui-image": "^3.3.2",
     "react": "^17.0.2",
@@ -43,6 +44,7 @@
     "redux-saga": "^1.1.3",
     "styled-components": "^5.3.3",
     "sweetalert2": "^11.1.10",
+    "typescript": "^4.5.2",
     "web-vitals": "^1.0.1"
   },
   "scripts": {
@@ -69,7 +71,6 @@
       "last 1 safari version"
     ]
   },
-
   "proxy": "http://218.154.207.223:8080",
   "devDependencies": {
     "customize-cra": "^1.0.0",
