+++ front/src/components/common/modules/ReqSwitch.js
