+++ front/src/components/common/modules/SurveyDialog.js
