+++ front/src/components/survey/comp/BeforeSurveyComp.js
@@ -1,4 +1,4 @@
-import React , { useEffect } from 'react';
+import React, { useEffect } from 'react';
 import { useSelector } from 'react-redux';
 import BeforeSurvey from '../UI/BeforeSurvey';
