+++ front/src/pages/ReadOnlyPage.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import Footer from '../components/common/UI/Footer';
-import Header from '../components/common/UI/Header';
+import Footer from '../components/common/modules/Footer';
+import Header from '../components/common/modules/Header';
 import SurveySubmitComp from '../components/survey/comp/SurveySubmitComp';
 
 const ReadOnlyPage = (props) => {
